requirements.txt
streamlit
pandas
sqlalchemy
pymysql
python-dotenv
tensorflow
scikit-learn
numpy